# Rework to get rid of concurrency issues