… deps/deps_test.go
… go.sum
… hugolib/cascade_test.go
… hugolib/content_render_hooks_test.go
… hugolib/hugo_modules_test.go
… hugolib/hugo_sites_build_errors_test.go
… hugolib/hugo_sites_rebuild_test.go
… hugolib/page_test.go
… hugolib/template_test.go
… hugolib/testhelpers_test.go
… output/layout.go
… output/layout_base.go
… output/layout_base_test.go
… output/layout_test.go
… resources/resource_transformers/templates/execute_as_template.go
… tpl/collections/apply_test.go
… tpl/template.go
… tpl/templates/templates.go
… tpl/tplimpl/shortcodes.go
… tpl/tplimpl/template.go
… tpl/tplimpl/templateProvider.go
… tpl/tplimpl/template_ast_transformers.go
… tpl/tplimpl/template_ast_transformers_test.go
… tpl/tplimpl/template_errors.go
… tpl/tplimpl/template_funcs_test.go
… tpl/tplimpl/template_info_test.go