This more or less completes the simplification of the template handling code in Hugo started in v0.62.

The main motivation was to fix a long lasting issue about a crash in HTML content files without front matter.

But this commit also comes with a big functional improvement.

As we now have moved the base template evaluation to the build stage we now use the same lookup rules for `baseof` as for `list` etc. type of templates.

This means that in this simple example you can have a `baseof` template for the `blog` section without having to duplicate the others::

```
layouts
├── _default
│   ├── baseof.html
│   ├── list.html
│   └── single.html
└── blog
    └── baseof.html
```

Also, when simplifying code, you often get rid of some double work, as shown in the "site building" benchmarks below.

These benchmarks looks suspiciously good:

```
name                              old time/op    new time/op    delta
SiteNew/Bundle_with_image-16        13.0ms ± 1%    10.5ms ± 1%  -19.55%  (p=0.029 n=4+4)
SiteNew/Bundle_with_JSON_file-16    13.1ms ± 1%    10.6ms ± 0%  -18.91%  (p=0.029 n=4+4)
SiteNew/Tags_and_categories-16      46.5ms ± 0%    42.8ms ± 1%   -7.96%  (p=0.029 n=4+4)
SiteNew/Canonify_URLs-16            52.3ms ±12%    47.9ms ± 5%     ~     (p=0.200 n=4+4)
SiteNew/Deep_content_tree-16        76.5ms ± 1%    70.2ms ± 1%   -8.27%  (p=0.029 n=4+4)
SiteNew/Many_HTML_templates-16      42.7ms ± 1%    36.8ms ± 1%  -13.93%  (p=0.029 n=4+4)
SiteNew/Page_collections-16         57.4ms ± 0%    52.4ms ± 0%   -8.64%  (p=0.029 n=4+4)

name                              old alloc/op   new alloc/op   delta
SiteNew/Bundle_with_image-16        3.81MB ± 0%    2.22MB ± 0%  -41.78%  (p=0.029 n=4+4)
SiteNew/Bundle_with_JSON_file-16    3.60MB ± 0%    2.00MB ± 0%  -44.34%  (p=0.029 n=4+4)
SiteNew/Tags_and_categories-16      19.1MB ± 2%    14.1MB ± 0%  -26.48%  (p=0.029 n=4+4)
SiteNew/Canonify_URLs-16            70.6MB ± 0%    68.9MB ± 0%   -2.42%  (p=0.029 n=4+4)
SiteNew/Deep_content_tree-16        37.0MB ± 0%    31.2MB ± 0%  -15.77%  (p=0.029 n=4+4)
SiteNew/Many_HTML_templates-16      17.5MB ± 0%    10.5MB ± 0%  -39.88%  (p=0.029 n=4+4)
SiteNew/Page_collections-16         25.9MB ± 0%    21.2MB ± 0%  -17.96%  (p=0.029 n=4+4)

name                              old allocs/op  new allocs/op  delta
SiteNew/Bundle_with_image-16         52.3k ± 0%     26.1k ± 0%  -50.17%  (p=0.029 n=4+4)
SiteNew/Bundle_with_JSON_file-16     52.3k ± 0%     26.1k ± 0%  -50.16%  (p=0.029 n=4+4)
SiteNew/Tags_and_categories-16        335k ± 1%      269k ± 0%  -19.53%  (p=0.029 n=4+4)
SiteNew/Canonify_URLs-16              422k ± 0%      395k ± 0%   -6.43%  (p=0.029 n=4+4)
SiteNew/Deep_content_tree-16          400k ± 0%      313k ± 0%  -21.66%  (p=0.029 n=4+4)
SiteNew/Many_HTML_templates-16        247k ± 0%      143k ± 0%  -42.05%  (p=0.029 n=4+4)
SiteNew/Page_collections-16           282k ± 0%      207k ± 0%  -26.46%  (p=0.029 n=4+4)
```

Fixes #6716
Fixes #6760